Ignore logs on first test
